+++ docs/changelog.rst
@@ -12,6 +12,19 @@ Changelog
 .. module:: joserfc
     :noindex:
 
+1.2.0
+-----
+
+**Unreleased**
+
+**Breaking changes**:
+
+- Enable "RFC7797" by default, use the ``joserfc.jws`` module directly.
+  - Use ``joserfc.jws.serialize_compact`` instead of ``joserfc.rfc7797.serialize_compact``
+  - Use ``joserfc.jws.deserialize_compact`` instead of ``joserfc.rfc7797.deserialize_compact``
+  - Use ``joserfc.jws.serialize_json`` instead of ``joserfc.rfc7797.serialize_json``
+  - Use ``joserfc.jws.deserialize_json`` instead of ``joserfc.rfc7797.deserialize_json``
+
 1.1.0
 -----